Refactor useStripePalette imports to use hooks directory and create useStripePalette hook
